requirements.txt
cec==0.2.8
pkg_resources==0.0.0
pycparser==2.21
pyserial==3.5
websocket-client==1.3.3
